Try move the Docker stuff into a separate package (for now)
