+++ example/requirements-test.txt
@@ -1,4 +1,4 @@
-docker
+-e docker-helper/
 iso8601
 requests
 testtools
