+++ example/setup.py
@@ -1,11 +1,11 @@
-from setuptools import setup, find_packages
+from setuptools import setup
 
 setup(
     name='mysite',
     version='0.1',
     author='Praekelt.org',
     author_email='sre@praekelt.org',
-    packages=find_packages(),
+    packages=['mysite'],
     install_requires=[
         'celery >=4.0, <4.1',
         'Django >=1.11, <1.12',
